stock_market_chat/requirements.txt
fastapi
uvicorn
httpx
sse-starlette
langchain
langchain-community
langchain-pinecone
langchain-nomic
pinecone-client
langchain-groq
finnhub-python 